+++ oap-server/server-query-plugin/query-graphql-plugin/src/main/java/org/apache/skywalking/oap/query/graphql/type/LogTestRequest.java
@@ -16,18 +16,15 @@
  *
  */
 
-package org.apache.skywalking.oap.query.graphql;
+package org.apache.skywalking.oap.query.graphql.type;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.apache.skywalking.oap.server.library.module.ModuleConfig;
 
-/**
- * The config of {@code query.graphql}.
- */
 @Getter
 @Setter
-public class GraphQLQueryConfig extends ModuleConfig {
-    private String path;
-    private boolean enableLogTestTool;
+public class LogTestRequest {
+    private String log;
+
+    private String dsl;
 }